requirements.txt
Flask==3.0.0
transformers>=4.38.0
torch>=2.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
redis>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
